requirements.txt
# DataPrep Pipeline — Dependencies
# Core
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2
requests==2.31.0

# Airflow (install separately if needed)
apache-airflow==2.9.1

# Dashboard
streamlit==1.35.0
plotly==5.22.0

# Testing
pytest==8.2.2

# Reporting
jinja2==3.1.4

# Data quality extras
scipy==1.13.1
